created reusable component
